@@ -16,7 +16,7 @@
         }
     ],
     "require": {
-        "php": "^7.3 || ^8.0",
+        "php": "^8.0",
         "illuminate/database": "^8.0 || ^9.0",
         "illuminate/http": "^8.0 || ^9.0"
     },
