@@ -163,8 +163,9 @@ trait FilterCreator
      */
     public static function boolean(string $property, callable $callback, ?callable $default = null): self
     {
-        return self::callback($property, static function (Builder $query, $value, $property) use ($callback, $default) {
-            return $query->when($value, $callback, $default);
-        });
+        return self::callback(
+            $property,
+            static fn (Builder $query, $value, $property): Builder => $query->when($value, $callback, $default)
+        );
     }
 }
