@@ -6,7 +6,7 @@ use Symplify\EasyCodingStandard\Config\ECSConfig;
 use Zing\CodingStandard\Set\ECSSetList;
 
 return static function (ECSConfig $ecsConfig): void {
-    $ecsConfig->sets([ECSSetList::PHP_73, ECSSetList::CUSTOM]);
+    $ecsConfig->sets([ECSSetList::PHP_80, ECSSetList::CUSTOM]);
     $ecsConfig->parallel();
     $ecsConfig->paths(
         [
