@@ -9,11 +9,9 @@ use Zing\QueryBuilder\Paginator;
 trait Pageable
 {
     /**
-     * @param string|\Zing\QueryBuilder\Paginator $paginator
-     *
      * @return $this
      */
-    public function enablePaginator($paginator = null)
+    public function enablePaginator(string|Paginator|null $paginator = null)
     {
         $paginator = $paginator instanceof Paginator ? $paginator : Paginator::name($paginator);
         $perPage = $this->request->input($paginator->getName()) ?: $paginator->getDefault();
