@@ -8,9 +8,9 @@ use Illuminate\Database\Eloquent\Builder;
 
 interface Filter
 {
-    /**
-     * @param mixed $value
-     * @param \Illuminate\Database\Query\Expression|string $property
-     */
-    public function apply(Builder $query, $value, $property): Builder;
+    public function apply(
+        Builder $query,
+        mixed $value,
+        \Illuminate\Database\Query\Expression|string $property
+    ): Builder;
 }
