@@ -22,7 +22,7 @@ class CallbackFilter implements Filter
         $this->callback = $callback;
     }
 
-    public function apply(Builder $query, $value, $property): Builder
+    public function apply(Builder $query, mixed $value, \Illuminate\Database\Query\Expression|string $property): Builder
     {
         \call_user_func($this->callback, $query, $value, $property);
 
